server_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
openai==1.3.8
python-dotenv==1.0.0
cachetools==5.3.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1